Extract search-highlight code from VirtualizedImageViewer
The overlay positioning maths, its types and its styled component made
up
roughly 40% of VirtualizedImageViewer.tsx but were only ever used by the
search-highlight feature, so they now live in
VirtualizedImageViewer.SearchHighlights.tsx.

The state and effect move too, behind a useSearchTermHighlights hook.
The
hook returns setImageRect/setImageContainerRect because the overlay
positions can only be calculated once IIIFItem has measured where the
image actually rendered - returning them makes that dependency explicit
rather than leaving it implied by state scattered through ItemRenderer.

Everything moved verbatim, including the effect's dependency array, so
there is no behaviour change. VirtualizedImageViewer.tsx drops from 436
to 236 lines.
